Moved animations to utilities and spinner functionality
