+++ jarviscli/plugins/football.py
@@ -2,7 +2,7 @@ import requests
 from tabulate import tabulate
 from colorama import Fore
 from plugin import plugin, require
-from plugins.animations import SpinnerThread
+from utilities.animations import SpinnerThread
 
 API_KEY = '1ebd3b92bf5041249f8c1e7a540ce98c'
 url = 'https://api.football-data.org/v2'
