+++ jarviscli/plugins/voice.py
@@ -4,17 +4,17 @@ from plugin import plugin
 @plugin('enable sound')
 def enable_sound(jarvis, s):
     """Let Jarvis use his voice."""
-    jarvis.enable_voice()
+    jarvis.enable_voice(jarvis)
 
 
 @plugin('disable sound')
-def disable_sound(jarvis, s):
+def disable_sound(s):
     """Deny Jarvis his voice."""
-    jarvis.disable_voice()
+    jarvis.disable_voice(jarvis)
 
 
 @plugin('say')
-def say(jarvis, s):
+def say(s):
     """Reads what is typed."""
     if not s:
         jarvis.say("What should I say?")
