+++ jarviscli/utilities/animations.py
